+++ crates/wdl-analysis/src/queue.rs
@@ -19,6 +19,7 @@ use lsp_types::CompletionResponse;
 use lsp_types::GotoDefinitionResponse;
 use lsp_types::Hover;
 use lsp_types::Location;
+use lsp_types::WorkspaceEdit;
 use parking_lot::RwLock;
 use petgraph::Direction;
 use petgraph::graph::NodeIndex;
@@ -75,6 +76,8 @@ pub enum Request<Context> {
     Completion(CompletionRequest<Context>),
     /// A request to get information about a symbol on hover.
     Hover(HoverRequest),
+    /// A request to rename a symbol workspace wide.
+    Rename(RenameRequest),
 }
 
 /// Represents a request to add documents to the graph.
@@ -185,6 +188,20 @@ pub struct HoverRequest {
     pub encoding: SourcePositionEncoding,
     /// The sender for completing the request.
     pub completed: oneshot::Sender<Option<Hover>>,
+}
+
+/// Represents a request to rename a symbol at a given position.
+pub struct RenameRequest {
+    /// The document where the request was initiated.
+    pub document: Url,
+    /// The position of the symbol in the document.
+    pub position: SourcePosition,
+    /// The encoding used for the position.
+    pub encoding: SourcePositionEncoding,
+    /// The new name of the symbol.
+    pub new_name: String,
+    /// The sender for completing the request.
+    pub completed: oneshot::Sender<Option<WorkspaceEdit>>,
 }
 
 /// A simple enumeration to signal a cancellation to the caller.
@@ -501,6 +518,7 @@ where
                     );
 
                     let graph = self.graph.read();
+
                     match handlers::hover(&graph, &document, position, encoding) {
                         Ok(result) => {
                             debug!(
@@ -512,6 +530,36 @@ where
                         }
                         Err(err) => {
                             debug!("error occurred while completing hover request: {err:?}");
+                            completed.send(None).ok();
+                        }
+                    }
+                }
+
+                Request::Rename(RenameRequest {
+                    document,
+                    position,
+                    encoding,
+                    new_name,
+                    completed,
+                }) => {
+                    let start = Instant::now();
+                    debug!(
+                        "received request for rename at {document}: {line}:{char}",
+                        line = position.line,
+                        char = position.character
+                    );
+
+                    let graph = self.graph.read();
+                    match handlers::rename(&graph, document, position, encoding, new_name) {
+                        Ok(result) => {
+                            debug!(
+                                "rename request completed in {elapsed:?}",
+                                elapsed = start.elapsed()
+                            );
+                            completed.send(result).ok();
+                        }
+                        Err(err) => {
+                            debug!("error occurred while completing rename request: {err:?}");
                             completed.send(None).ok();
                         }
                     }