+++ crates/wdl-lsp/src/server.rs
@@ -417,6 +417,7 @@ impl LanguageServer for Server {
                     ..Default::default()
                 }),
                 hover_provider: Some(HoverProviderCapability::Simple(true)),
+                rename_provider: Some(OneOf::Left(true)),
                 ..Default::default()
             },
             server_info: Some(ServerInfo {
@@ -826,6 +827,33 @@ impl LanguageServer for Server {
                 params.text_document_position_params.text_document.uri,
                 position,
                 SourcePositionEncoding::UTF16,
+            )
+            .await
+            .map_err(|e| RpcError {
+                code: ErrorCode::InternalError,
+                message: e.to_string().into(),
+                data: None,
+            })?;
+        Ok(result)
+    }
+
+    async fn rename(&self, mut params: RenameParams) -> RpcResult<Option<WorkspaceEdit>> {
+        normalize_uri_path(&mut params.text_document_position.text_document.uri);
+
+        debug!("received `textDocument/rename` request: {params:#?}");
+
+        let position = SourcePosition::new(
+            params.text_document_position.position.line,
+            params.text_document_position.position.character,
+        );
+
+        let result = self
+            .analyzer
+            .rename(
+                params.text_document_position.text_document.uri,
+                position,
+                SourcePositionEncoding::UTF16,
+                params.new_name,
             )
             .await
             .map_err(|e| RpcError {