+++ crates/wdl-analysis/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* Added rename support for the WDL Language Server ([#563](https://github.com/stjude-rust-labs/wdl/pull/563)).
 * Added hover support for the WDL Language Server ([#540](https://github.com/stjude-rust-labs/wdl/pull/540)).
 
 ## 0.11.0 - 07-31-2025